requirements.txt
streamlit
Pillow
torch
transformers
numpy
opencv-python
